REWORD ONLY SQUASH
strbuf: change an always NULL/"" strbuf_addftime() param to bool

Change the code for deciding what's to be done about %Z to stop
passing either a NULL or "" char * to strbuf_addftime(). Instead
pass a boolean int to indicate whether the strftime() %Z format
should be suppressed by converting it to an empty string, which is
the best this code can when dealing with non-local timestamp to
be comformant to POSIX.

This code grew organically between the changes in 9eafe86d58 ("Merge
branch 'rs/strbuf-addftime-zZ'", 2017-06-22). The intent was to use
this API in the future to leave the door open to pass a custom
timezone name to the function (see [1] and related messages).

strbuf_addftime() allows callers to pass a time zone name for
expanding%Z.  The only current caller either passes the empty string
or NULL, in which case%Z is handed over verbatim to strftime(3).
Replace that string parameter with a flag controlling whether to
remove%Z from the format specification.  This simplifies the code.

1. https://public-inbox.org/git/CACBZZX5OQc45fUyDVayE89rkT=+8m5S4efSXCAbCy7Upme5zLA@mail.gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>