t0021: write "OUT <size>" only on success
"rot13-filter.pl" always writes "OUT <size>" to the debug log at the end
of a response.

This works without issues for the existing responses "abort",
"error", and "success". A new response "delayed", that will be
introduced in a subsequent patch, accepts the input without giving
the filtered result right away.  Since we cannot know the size of
the response at this point, we cannot write "OUT <size>" for
"delayed" responses.

To simplify the code we do not write "OUT <size>" for "abort" and
"error" responses, too, as their size is always zero.

Signed-off-by: Lars Schneider <larsxschneider@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>